refactor(evaluations): split auto-population helpers out of tasks.py
Move the seven auto-population functions and the Celery periodic task into
a new apps/evaluations/auto_population.py module, update the SCHEDULED_TASKS
entry in settings.py to the new task path, and point test imports accordingly.
